Use INTEGER instead of REAL type for sql sizes (#613)

+++ internal/migrations/000012_init-files-table.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS files (
   "id" TEXT NOT NULL,
-  "size" REAL NOT NULL,
+  "size" INTEGER NOT NULL,
   "mimetype" TEXT DEFAULT NULL,
   "checksum" TEXT NOT NULL,
   "key" BLOB NOT NULL,
